* Added some content to the repository README.txt * Added a x64 platform configuration to the Visual Studio solution * Commented out the test sleeping calls every frame * Made configuration fields print with PrintVerbose instead of printf * Commented out the fake test server entry * Added a daemonize() function for POSIX platforms to run as a daemon * PRINT_ERROR and PRINT_WARNING prints now go to stderr instead of stdout * Started new configuration handling stuff (config.h) * Playing around with WSA Events

+++ src/oob_cmds.cpp
@@ -102,7 +102,7 @@ void JAMS_GetServers(Command *cmd, NetAdr *from, Q3OobMsg *msg)
 		if (currentAdrs > 0)
 		{
 			char replyChunk[1024];
-			int writeCount = 0;
+			size_t writeCount = 0;
 
 			Strncpyz(replyChunk, "ÿÿÿÿgetserversResponse\n", sizeof(replyChunk));
 			writeCount = strlen(replyChunk);
